chore: Introduce domain model for player data
Convert to and from this when reading from hypixel and writing to http
responses.

We will start converting all ports+adapters that read/write player data
to read/write to/from this struct.

+++ internal/processing/parsing.go
@@ -16,8 +16,8 @@ type hypixelAPIResponse struct {
 type HypixelAPIPlayer struct {
 	UUID        *string          `json:"uuid,omitempty"`
 	Displayname *string          `json:"displayname,omitempty"`
-	LastLogin   *int             `json:"lastLogin,omitempty"`
-	LastLogout  *int             `json:"lastLogout,omitempty"`
+	LastLogin   *int64           `json:"lastLogin,omitempty"`
+	LastLogout  *int64           `json:"lastLogout,omitempty"`
 	Stats       *HypixelAPIStats `json:"stats,omitempty"`
 }
